fix: fix data type of price

+++ app/src/main/java/com/example/androidmycoffee/domain/model/Coffee.kt
@@ -1,10 +1,12 @@
 package com.example.androidmycoffee.domain.model
 
+import java.math.BigDecimal
+
 data class Coffee(
     val id: Int,
     val name: String,
     val type: TypeCoffee,
-    val price: Int,
+    val price: BigDecimal,
 )
 
 enum class TypeCoffee {
